Only clean up dev reg when options change
