+++ homeassistant/components/utility_meter/__init__.py
@@ -185,11 +185,10 @@ async def async_setup(hass: HomeAssistant, config: ConfigType) -> bool:
 
 async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     """Set up Utility Meter from a config entry."""
-
-    await async_remove_device_old(hass, entry)
-
     entity_registry = er.async_get(hass)
-    hass.data[DATA_UTILITY][entry.entry_id] = {}
+    hass.data[DATA_UTILITY][entry.entry_id] = {
+        "source": entry.options[CONF_SOURCE_SENSOR],
+    }
     hass.data[DATA_UTILITY][entry.entry_id][DATA_TARIFF_SENSORS] = []
 
     try:
@@ -225,7 +224,22 @@ async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
 
 async def config_entry_update_listener(hass: HomeAssistant, entry: ConfigEntry) -> None:
     """Update listener, called when the config entry options are changed."""
+    old_source = hass.data[DATA_UTILITY][entry.entry_id]["source"]
     await hass.config_entries.async_reload(entry.entry_id)
+
+    if old_source == entry.options[CONF_SOURCE_SENSOR]:
+        return
+
+    entity_registry = er.async_get(hass)
+    device_registry = dr.async_get(hass)
+
+    old_source_entity = entity_registry.async_get(old_source)
+    if not old_source_entity or not old_source_entity.device_id:
+        return
+
+    device_registry.async_update_device(
+        old_source_entity.device_id, remove_config_entry_id=entry.entry_id
+    )
 
 
 async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
@@ -256,22 +270,3 @@ async def async_migrate_entry(hass: HomeAssistant, config_entry: ConfigEntry) ->
     _LOGGER.info("Migration to version %s successful", config_entry.version)
 
     return True
-
-
-async def async_remove_device_old(hass: HomeAssistant, entry: ConfigEntry):
-    """Remove obsolete source entity devices."""
-    entity_registry = er.async_get(hass)
-    device_registry = dr.async_get(hass)
-
-    source_entity = entity_registry.async_get(entry.options[CONF_SOURCE_SENSOR])
-    source_device_id = None
-    if source_entity is not None and source_entity.device_id is not None:
-        source_device_id = source_entity.device_id
-
-    for device_entry in dr.async_entries_for_config_entry(
-        device_registry, entry.entry_id
-    ):
-        if device_entry.id != source_device_id:
-            device_registry.async_update_device(
-                device_entry.id, remove_config_entry_id=entry.entry_id
-            )
